Make the async search field generic and apply it to both company and team fields
